Makes dirs instead of downloading; chmod new files

+++ misc/dropbox-sync.py
@@ -3,39 +3,31 @@ import json
 import os
 import requests
 import shutil
+import sys
 import zipfile
 
 BASE_DIR = os.path.dirname(os.path.realpath(__file__))
 FILE_DIR = f'{BASE_DIR}/file_data'
 ZIP_FILE= f'{BASE_DIR}/files.zip'
 CURSOR_FILE = f'{BASE_DIR}/cursor'
+LOG_FILE = f'{BASE_DIR}/dropbox-sync.log'
 
 DROPBOX_FILE_DIR='/colrc-v2-files'
 DROPBOX_ACCESS_TOKEN='H66L2R8JfX8AAAAAAAAAASyxUvZolCy956ASqQZjEPhFHXDbfa5zbk1pmNdsxYSA'
 
-# TODO: Split download_folder so that print statements don't appear in the middle of update() output
-def download_folder(folder, dl_location, rename=False, progress=False):
+def download_all(dl_location):
     """
-    Downloads a folder from Dropbox as a .zip file. 
+    Downloads everything at DROPBOX_FILE_DIR as a .zip file. 
     
     Parameters
     ----------
-    folder : str or int
-        The folder's path or ID in Dropbox
-    
     dl_location : str
-        The path to extract the zip file to on the local filesystem (excluding the name of the folder)
-
-    rename : bool, default=True
-        Internal use only; renames the extracted folder from the Dropbox app's root directory to "file_data"
-
-    progress : bool, default=False
-        Dynamically prints the download's progress in the terminal (isn't really necessary for single folders)
+        The path to extract the zip file to on the local filesystem (excluding the name of the folder, since it will be named FILE_DIR).
     """
 
     headers = {
         'Authorization': f'Bearer {DROPBOX_ACCESS_TOKEN}',
-        'Dropbox-API-Arg': f'{{"path": "{folder}"}}'
+        'Dropbox-API-Arg': f'{{"path": "{DROPBOX_FILE_DIR}"}}'
     }
     
     response = requests.post('https://content.dropboxapi.com/2/files/download_zip', headers=headers, stream=True)
@@ -46,15 +38,13 @@ def download_folder(folder, dl_location, rename=False, progress=False):
         for chunk in response.iter_content(chunk_size=1024):
             f.write(chunk)
             megabytes_written += 1/1024
-            if progress:
-                print(f'\r{megabytes_written:.0f}MB downloaded...', end='', flush=True)
+            print(f'\r{megabytes_written:.0f}MB downloaded...', end='', flush=True)
     
-    print('\nExtracting zip file...')
+    print('Extracting zip file...')
     with zipfile.ZipFile(ZIP_FILE, 'r') as z:
         z.filename = 'file_data'
         z.extractall(dl_location)
 
-    if rename:
         print('Renaming folder...')
         os.rename(f'{BASE_DIR}/{DROPBOX_FILE_DIR}', FILE_DIR)
 
@@ -68,17 +58,19 @@ def download_folder(folder, dl_location, rename=False, progress=False):
         for f in files:
             os.chmod(os.path.join(root, f), 0o777)
 
+    save_cursor()
+
 def download_file(file, dl_location):
     """
-    Downloads a folder from Dropbox as a .zip file. 
+    Downloads a file from Dropbox. 
     
     Parameters
     ----------
     file : str or int
-        The folder's path or ID in Dropbox
+        The file's path or ID in Dropbox.
     
     dl_location : str
-        The path to download the file to on the local filesystem (excluding the name of the folder)
+        The path to download the file to on the local filesystem.
     """
 
     headers = {
@@ -88,6 +80,9 @@ def download_file(file, dl_location):
     
     response = requests.post('https://content.dropboxapi.com/2/files/download', headers=headers)
     
+    if os.path.isdir(dl_location):
+        raise FileExistsError
+
     with open(dl_location, 'wb') as f:
         f.write(response.content)
             
@@ -125,44 +120,64 @@ def update(cursor):
     }
     data = f'{{"cursor": "{cursor}"}}'
 
-    updated_files = json.loads(requests.post('https://api.dropboxapi.com/2/files/list_folder/continue', headers=headers, data=data).content)
+    response = requests.post('https://api.dropboxapi.com/2/files/list_folder/continue', headers=headers, data=data)
+    updated_files = json.loads(response.content)
     
-    print('\nUpdating files:')
-    for entry in updated_files['entries']:
-        local_path = f'{FILE_DIR}{entry["path_display"].replace(DROPBOX_FILE_DIR, "")}'
+    if updated_files['entries']:
+        print('\nUpdating your current directory...')
+        for entry in updated_files['entries']:
+            local_path = f'{FILE_DIR}{entry["path_display"].replace(DROPBOX_FILE_DIR, "")}'
+            
+            if entry['.tag'] == 'deleted':
+                try:
+                    if os.path.isdir(local_path):
+                        shutil.rmtree(local_path)
+                        print(f'DEL {entry["name"]}')
+                    elif os.path.isfile(local_path):
+                        os.remove(local_path)
+                        print(f'DEL {entry["name"]}')
+                    else:
+                        print(f'DEX {entry["name"]}')
+                except:
+                    print(f'ERR DEL {entry["name"]}')
+            
+            elif entry['.tag'] == 'file':
+                try:
+                    download_file(entry['path_display'], local_path)
+                    os.chmod(local_path, 0o777)
+                    print(f'ADD {entry["name"]}')
+                except FileExistsError:
+                    print(f'AEX {entry["name"]}')
+                except:
+                    print(f'ERR ADD {entry["name"]}')
+            
+            elif entry['.tag'] == 'folder':
+                try:
+                    os.mkdir(local_path)
+                    os.chmod(local_path, 0o777)
+                    print(f'ADD {entry["name"]}')
+                except FileExistsError:
+                    print(f'AEX {entry["name"]}')
+                except:
+                    print(f'ERR ADD {entry["name"]}')
+            
+            else:
+                print(f'Unknown error has occurred. Please check the log (located at {LOG_FILE}) for details.')
+                with open(LOG_FILE, 'w') as f:
+                    f.write(response.content)
+                sys.exit(1)
         
-        if entry['.tag'] == 'deleted':
-            try:
-                if os.path.isdir(local_path):
-                    shutil.rmtree(local_path)
-                else:
-                    os.remove(local_path)
-                print(f'DEL {entry["name"]}')
-            except:
-                print(f'ERR DEL {entry["name"]}')
-        
-        elif entry['.tag'] == 'file':
-            try:
-                download_file(entry['path_display'], local_path)
-                print(f'ADD {entry["name"]}')
-            except:
-                print(f'ERR ADD {entry["name"]}')
-        
-        elif entry['.tag'] == 'folder':
-            try:
-                download_folder(entry['path_display'], local_path.replace(entry['name'], ''))
-                print(f'ADD {entry["name"]}')
-            except:
-                print(f'ERR ADD {entry["name"]}')
+        save_cursor()
+    
+    else:
+        print('\nNo files to update.')
 
 # TODO: Support longer output from Dropbox, which comes with its own cursor(s) 
 def main():
     if os.path.isdir(FILE_DIR):
         if os.path.isfile(CURSOR_FILE):
-            print('\nUpdating your current directory...')
             with open(CURSOR_FILE) as f:
                 update(f.readline())
-                save_cursor()
         
         else:
             print('\nWhile an nginx file directory was found, a Dropbox cursor was not.')
@@ -171,12 +186,10 @@ def main():
             
             if click.confirm('Would you like to automatically remove and recreate the folder with a new cursor?'):
                 shutil.rmtree(FILE_DIR)
-                download_folder(DROPBOX_FILE_DIR, BASE_DIR, rename=True, progress=True)
-                save_cursor()
+                download_all(BASE_DIR)
     
     else:
         print('\nNo nginx file directory found.')
-        download_folder(DROPBOX_FILE_DIR, BASE_DIR, rename=True, progress=True)
-        save_cursor()
+        download_all(BASE_DIR)
 
 main()
